Move tab to sub-component
